% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectralmetalearning.R
\name{eigensml}
\alias{eigensml}
\title{Take the SML prediction of multiple predictions using the eigenvectors method.}
\usage{
eigensml(predictors)
}
\arguments{
\item{predictors}{A matrix of predictors, with the group as the row, and predictions as the columns}
}
\value{
The SML combined prediction of the predictors
}
\description{
Take the SML prediction of multiple predictions using the eigenvectors method.
}
\examples{
\dontrun{prediction = eigensml(multiplePredictions)}
}
\seealso{
\code{\link{svdsml}}
}

