% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectralmetalearning.R
\name{plotaccuracies}
\alias{plotaccuracies}
\title{Plot the balanced accuracies of the predictors, as estimated by the SML}
\usage{
plotaccuracies(predictors)
}
\arguments{
\item{predictors}{A matrix of predictors, with the group as the row, and predictions as the columns}
}
\value{
A plot of the predictors accuracies
}
\description{
Plot the balanced accuracies of the predictors, as estimated by the SML
}
\examples{
\dontrun{plotaccuracies(multiplePredictions)}
}

